+++ lib/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
+var utilAddress = require('./util/splitAddress');
 /**
  * From jquery.Thailand.js line 38 - 100
  */
-
 var preprocess = function preprocess(data) {
   var lookup = [];
   var words = [];
@@ -107,69 +107,15 @@ exports.searchAddressByZipcode = function (searchStr, maxResult) {
 
 exports.splitAddress = function (fullAddress) {
   var regex = /\s(\d{5})(\s|$)/gi;
-  var zip = regex.exec(fullAddress)[1];
-  // let zip = result
-  var address = fullAddress;
-  address = address.replace(zip, '');
-  address = address.replace('Thailand', '');
-  address = address.replace('ต.', '');
-  address = address.replace('อ.', '');
-  address = address.replace('จ.', '');
-  address = address.replace('ตำบล', '');
-  address = address.replace('อำเภอ', '');
-  address = address.replace('จังหวัด', '');
-  address = address.replace('แขวง', '');
-  address = address.replace('เขต', '');
-  address = address.replace('แขวง.', '');
-  address = address.replace('เขต.', '');
-  address = address.replace(' กทม. ', ' กรุงเทพมหานคร ');
-  address = address.replace(' กทม ', ' กรุงเทพมหานคร ');
-  address = address.replace(' กรุงเทพ ', ' กรุงเทพมหานคร ');
-  // console.log(address)
-  var searchResult = exports.searchAddressByZipcode(zip);
-  // console.log(searchResult)
-  searchResult.forEach(function (element, index) {
-    var district = address.indexOf(element.district);
-    var next = district !== -1 ? district + 1 : 0;
-    var amphoe = address.indexOf(element.amphoe, next);
-    next = amphoe !== -1 ? amphoe + 1 : next + 1;
-    var province = address.indexOf(element.province, next);
-    // console.log(district, amphoe, province)
-    var point = [district, amphoe, province].filter(function (el) {
-      return el >= 0;
-    }).length;
-    // console.log(point)
-    searchResult[index].point = point;
-  });
-  searchResult.sort(function (a, b) {
-    return b.point - a.point;
-  });
-  var result = searchResult[0];
+  var regexResult = regex.exec(fullAddress);
+  if (!regexResult) {
+    return null;
+  }
+  var zip = regexResult[1];
+  var address = utilAddress.prepareAddress(fullAddress, zip);
+  var result = utilAddress.getBestResult(zip, address);
   if (result) {
-    // console.log(result)
-    var newAddress = address;
-
-    var regexDistrict = new RegExp('\\s' + result.district, 'g');
-    var findDistrict = regexDistrict.exec(newAddress);
-    if (findDistrict) {
-      newAddress = newAddress.replace(findDistrict[0], '');
-    }
-
-    var regexAmphoe = new RegExp('\\s' + result.amphoe + '|\u0E40\u0E21\u0E37\u0E2D\u0E07', 'g');
-    var findAmphoe = regexAmphoe.exec(newAddress);
-    if (findAmphoe) {
-      newAddress = newAddress.replace(findAmphoe[0], '');
-    }
-
-    var regexProvince = new RegExp('\\s' + result.province, 'g');
-    var findProvince = regexProvince.exec(newAddress);
-    if (findProvince) {
-      newAddress = newAddress.replace(findProvince[0], '');
-    }
-
-    newAddress = newAddress.trim();
-    // console.log(newAddress)
-
+    var newAddress = utilAddress.cleanupAddress(address, result);
     return {
       address: newAddress,
       district: result.district,
@@ -178,11 +124,5 @@ exports.splitAddress = function (fullAddress) {
       zipcode: zip
     };
   }
-  return {
-    address: address,
-    district: '',
-    amphoe: '',
-    province: '',
-    zipcode: zip
-  };
+  return null;
 };
