+++ npm-debug.log
@@ -1,0 +1,30 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/Users/n9ti/.nvm/versions/node/v6.11.3/bin/node',
+1 verbose cli   '/Users/n9ti/.nvm/versions/node/v6.11.3/bin/npm',
+1 verbose cli   'version',
+1 verbose cli   'patch' ]
+2 info using npm@3.10.10
+3 info using node@v6.11.3
+4 info git [ 'status', '--porcelain' ]
+5 verbose stack Error: Git working directory not clean.
+5 verbose stack M lib/index.js
+5 verbose stack     at /Users/n9ti/.nvm/versions/node/v6.11.3/lib/node_modules/npm/lib/version.js:247:19
+5 verbose stack     at /Users/n9ti/.nvm/versions/node/v6.11.3/lib/node_modules/npm/lib/utils/no-progress-while-running.js:21:8
+5 verbose stack     at ChildProcess.exithandler (child_process.js:189:7)
+5 verbose stack     at emitTwo (events.js:106:13)
+5 verbose stack     at ChildProcess.emit (events.js:191:7)
+5 verbose stack     at maybeClose (internal/child_process.js:920:16)
+5 verbose stack     at Socket.<anonymous> (internal/child_process.js:351:11)
+5 verbose stack     at emitOne (events.js:96:13)
+5 verbose stack     at Socket.emit (events.js:188:7)
+5 verbose stack     at Pipe._handle.close [as _onclose] (net.js:497:12)
+6 verbose cwd /Users/n9ti/Working/thai-address-database
+7 error Darwin 17.2.0
+8 error argv "/Users/n9ti/.nvm/versions/node/v6.11.3/bin/node" "/Users/n9ti/.nvm/versions/node/v6.11.3/bin/npm" "version" "patch"
+9 error node v6.11.3
+10 error npm  v3.10.10
+11 error Git working directory not clean.
+11 error M lib/index.js
+12 error If you need help, you may report this error at:
+12 error     <https://github.com/npm/npm/issues>
+13 verbose exit [ 1, true ]
